use a generic file with the argt class name being a macro
since several free functions needs to be defined, it seems
to be the simplest and cleanest solution
